+++ nautilus_core/model/src/identifiers/symbol.rs
@@ -13,19 +13,19 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use nautilus_core::buffer::{Buffer, Buffer32};
+use nautilus_core::buffer::{Buffer, Buffer128};
 use std::fmt::{Debug, Display, Formatter, Result};
 
 #[repr(C)]
 #[derive(Clone, Hash, PartialEq, Debug)]
 pub struct Symbol {
-    pub value: Buffer32,
+    pub value: Buffer128, // TODO: Temporary to support Betfair
 }
 
 impl From<&str> for Symbol {
     fn from(s: &str) -> Symbol {
         Symbol {
-            value: Buffer32::from(s),
+            value: Buffer128::from(s),
         }
     }
 }
@@ -45,12 +45,12 @@ pub extern "C" fn symbol_free(symbol: Symbol) {
 }
 
 #[no_mangle]
-pub extern "C" fn symbol_from_bytes(value: Buffer32) -> Symbol {
+pub extern "C" fn symbol_from_bytes(value: Buffer128) -> Symbol {
     Symbol { value }
 }
 
 #[no_mangle]
-pub extern "C" fn symbol_to_bytes(symbol: Symbol) -> Buffer32 {
+pub extern "C" fn symbol_to_bytes(symbol: Symbol) -> Buffer128 {
     symbol.value
 }