+++ nautilus_core/model/src/orderbook/level.rs
@@ -164,8 +164,8 @@ mod tests {
 
     #[test]
     fn test_level_comparisons_bid_side() {
-        let level0 = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::BUY));
-        let level1 = Level::new(BookPrice::new(Price::new(1.01, 2), OrderSide::BUY));
+        let level0 = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::Buy));
+        let level1 = Level::new(BookPrice::new(Price::new(1.01, 2), OrderSide::Buy));
 
         assert_eq!(level0, level0);
         assert!(level0 > level1);
@@ -173,8 +173,8 @@ mod tests {
 
     #[test]
     fn test_level_comparisons_ask_side() {
-        let level0 = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::SELL));
-        let level1 = Level::new(BookPrice::new(Price::new(1.01, 2), OrderSide::SELL));
+        let level0 = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::Sell));
+        let level1 = Level::new(BookPrice::new(Price::new(1.01, 2), OrderSide::Sell));
 
         assert_eq!(level0, level0);
         assert!(level0 < level1);
@@ -182,11 +182,11 @@ mod tests {
 
     #[test]
     fn test_level_add_one_order() {
-        let mut level = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::BUY));
+        let mut level = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::Buy));
         let order = Order::new(
             Price::new(1.00, 2),
             Quantity::new(10.0, 0),
-            OrderSide::BUY,
+            OrderSide::Buy,
             0,
         );
 
@@ -199,17 +199,17 @@ mod tests {
 
     #[test]
     fn test_level_add_multiple_orders() {
-        let mut level = Level::new(BookPrice::new(Price::new(2.00, 2), OrderSide::BUY));
+        let mut level = Level::new(BookPrice::new(Price::new(2.00, 2), OrderSide::Buy));
         let order1 = Order::new(
             Price::new(2.00, 2),
             Quantity::new(10.0, 0),
-            OrderSide::BUY,
+            OrderSide::Buy,
             0,
         );
         let order2 = Order::new(
             Price::new(2.00, 2),
             Quantity::new(20.0, 0),
-            OrderSide::BUY,
+            OrderSide::Buy,
             1,
         );
 
@@ -223,17 +223,17 @@ mod tests {
 
     #[test]
     fn test_level_update_order() {
-        let mut level = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::BUY));
+        let mut level = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::Buy));
         let order1 = Order::new(
             Price::new(1.00, 2),
             Quantity::new(10.0, 0),
-            OrderSide::BUY,
+            OrderSide::Buy,
             0,
         );
         let order2 = Order::new(
             Price::new(1.00, 2),
             Quantity::new(20.0, 0),
-            OrderSide::BUY,
+            OrderSide::Buy,
             0,
         );
 
@@ -247,17 +247,17 @@ mod tests {
 
     #[test]
     fn test_level_update_order_with_zero_size_deletes() {
-        let mut level = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::BUY));
+        let mut level = Level::new(BookPrice::new(Price::new(1.00, 2), OrderSide::Buy));
         let order1 = Order::new(
             Price::new(1.00, 2),
             Quantity::new(10.0, 0),
-            OrderSide::BUY,
+            OrderSide::Buy,
             0,
         );
         let order2 = Order::new(
             Price::new(1.00, 2),
             Quantity::new(0.0, 0),
-            OrderSide::BUY,
+            OrderSide::Buy,
             0,
         );