+++ nautilus_trader/model/objects.pyx
@@ -34,13 +34,13 @@ from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.rust.model cimport FIXED_SCALAR
 from nautilus_trader.core.rust.model cimport Currency_t
 from nautilus_trader.core.rust.model cimport money_free
-from nautilus_trader.core.rust.model cimport money_from_fixed
+from nautilus_trader.core.rust.model cimport money_from_raw
 from nautilus_trader.core.rust.model cimport money_new
 from nautilus_trader.core.rust.model cimport price_free
-from nautilus_trader.core.rust.model cimport price_from_fixed
+from nautilus_trader.core.rust.model cimport price_from_raw
 from nautilus_trader.core.rust.model cimport price_new
 from nautilus_trader.core.rust.model cimport quantity_free
-from nautilus_trader.core.rust.model cimport quantity_from_fixed
+from nautilus_trader.core.rust.model cimport quantity_from_raw
 from nautilus_trader.core.rust.model cimport quantity_new
 from nautilus_trader.core.string cimport precision_from_str
 from nautilus_trader.model.currency cimport Currency
@@ -83,13 +83,13 @@ cdef class Quantity:
         Condition.true(precision <= 9, "invalid precision, was > 9")
         Condition.true(value >= 0.0, f"quantity negative, was {value}")
 
-        self._qty = quantity_new(value, precision)
+        self._mem = quantity_new(value, precision)
 
     def __getstate__(self):
-        return self._qty.fixed, self._qty.precision
+        return self._mem.raw, self._mem.precision
 
     def __setstate__(self, state):
-        self._qty = quantity_from_fixed(state[0], state[1])
+        self._mem = quantity_from_raw(state[0], state[1])
 
     def __eq__(self, other) -> bool:
         return Quantity._compare(self, other, Py_EQ)
@@ -185,60 +185,60 @@ cdef class Quantity:
         return int(self.as_f64_c())
 
     def __hash__(self) -> int:
-        return hash(self._qty.fixed)
+        return hash(self._mem.raw)
 
     def __str__(self) -> str:
-        return f"{self.as_f64_c():.{self._qty.precision}f}"
+        return f"{self.as_f64_c():.{self._mem.precision}f}"
 
     def __repr__(self) -> str:
         return f"{type(self).__name__}('{self}')"
 
     def __del__(self) -> None:
         # https://cython.readthedocs.io/en/latest/src/userguide/special_methods.html#finalization-methods-dealloc-and-del
-        quantity_free(self._qty)  # `self._qty` moved to Rust (then dropped)
+        quantity_free(self._mem)  # `self._mem` moved to Rust (then dropped)
 
     cdef bint eq(self, Quantity other) except *:
-        return self._qty.fixed == other._qty.fixed
+        return self._mem.raw == other._mem.raw
 
     cdef bint ne(self, Quantity other) except *:
-        return self._qty.fixed != other._qty.fixed
+        return self._mem.raw != other._mem.raw
 
     cdef bint lt(self, Quantity other) except *:
-        return self._qty.fixed < other._qty.fixed
+        return self._mem.raw < other._mem.raw
 
     cdef bint le(self, Quantity other) except *:
-        return self._qty.fixed <= other._qty.fixed
+        return self._mem.raw <= other._mem.raw
 
     cdef bint gt(self, Quantity other) except *:
-        return self._qty.fixed > other._qty.fixed
+        return self._mem.raw > other._mem.raw
 
     cdef bint ge(self, Quantity other) except *:
-        return self._qty.fixed >= other._qty.fixed
+        return self._mem.raw >= other._mem.raw
 
     cdef bint is_zero(self) except *:
-        return self._qty.fixed == 0
+        return self._mem.raw == 0
 
     cdef bint is_negative(self) except *:
-        return self._qty.fixed < 0
+        return self._mem.raw < 0
 
     cdef bint is_positive(self) except *:
-        return self._qty.fixed > 0
+        return self._mem.raw > 0
 
     @staticmethod
-    def from_fixed(uint64_t fixed, uint8_t precision):
-        return Quantity.from_fixed_c(fixed, precision)
+    def from_raw(uint64_t raw, uint8_t precision):
+        return Quantity.from_raw_c(raw, precision)
 
     @staticmethod
-    cdef Quantity from_fixed_c(uint64_t fixed, uint8_t precision):
+    cdef Quantity from_raw_c(uint64_t raw, uint8_t precision):
         cdef Quantity quantity = Quantity.__new__(Quantity)
-        quantity._qty = quantity_from_fixed(fixed, precision)
+        quantity._mem = quantity_from_raw(raw, precision)
         return quantity
 
-    cdef uint64_t fixed_uint64_c(self) except *:
-        return self._qty.fixed
+    cdef uint64_t raw_uint64_c(self) except *:
+        return self._mem.raw
 
     cdef double as_f64_c(self) except *:
-        return self._qty.fixed / FIXED_SCALAR
+        return self._mem.raw / FIXED_SCALAR
 
     @staticmethod
     cdef object _extract_decimal(object obj):
@@ -264,7 +264,7 @@ cdef class Quantity:
 
     @property
     def precision(self) -> int:
-        return self._qty.precision
+        return self._mem.precision
 
     @staticmethod
     cdef Quantity zero_c(uint8_t precision):
@@ -363,25 +363,25 @@ cdef class Quantity:
         str
 
         """
-        return f"{self.as_f64_c():,.{self._qty.precision}f}".replace(",", "_")
+        return f"{self.as_f64_c():,.{self._mem.precision}f}".replace(",", "_")
 
     cdef Quantity add(self, Quantity other):
-        cdef int64_t fixed = self._qty.fixed + other.fixed_int64_c()
-        return Quantity.from_fixed_c(fixed, self._qty.precision)
+        cdef int64_t raw = self._mem.raw + other.raw_int64_c()
+        return Quantity.from_raw_c(raw, self._mem.precision)
 
     cdef Quantity sub(self, Quantity other):
-        cdef int64_t fixed = self._qty.fixed - other.fixed_int64_c()
-        return Quantity.from_fixed_c(fixed, self._qty.precision)
+        cdef int64_t raw = self._mem.raw - other.raw_int64_c()
+        return Quantity.from_raw_c(raw, self._mem.precision)
 
     cdef void add_assign(self, Quantity other) except *:
-        self._qty.fixed += other.fixed_uint64_c()
-        if self._qty.precision == 0:
-            self._qty.precision = other.precision
+        self._mem.raw += other.raw_uint64_c()
+        if self._mem.precision == 0:
+            self._mem.precision = other.precision
 
     cdef void sub_assign(self, Quantity other) except *:
-        self._qty.fixed -= other.fixed_uint64_c()
-        if self._qty.precision == 0:
-            self._qty.precision = other.precision
+        self._mem.raw -= other.raw_uint64_c()
+        if self._mem.precision == 0:
+            self._mem.precision = other.precision
 
     cpdef object as_decimal(self):
         """
@@ -392,7 +392,7 @@ cdef class Quantity:
         Decimal
 
         """
-        return decimal.Decimal(f"{self.as_double():.{self._qty.precision}f}")
+        return decimal.Decimal(f"{self.as_double():.{self._mem.precision}f}")
 
     cpdef double as_double(self) except *:
         """
@@ -438,13 +438,13 @@ cdef class Price:
     def __init__(self, double value, uint8_t precision):
         Condition.true(precision <= 9, "invalid precision, was > 9")
 
-        self._price = price_new(value, precision)
+        self._mem = price_new(value, precision)
 
     def __getstate__(self):
-        return self._price.fixed, self._price.precision
+        return self._mem.raw, self._mem.precision
 
     def __setstate__(self, state):
-        self._price = price_from_fixed(state[0], state[1])
+        self._mem = price_from_raw(state[0], state[1])
 
     def __eq__(self, other) -> bool:
         return Price._compare(self, other, Py_EQ)
@@ -540,60 +540,60 @@ cdef class Price:
         return int(self.as_f64_c())
 
     def __hash__(self) -> int:
-        return hash(self._price.fixed)
+        return hash(self._mem.raw)
 
     def __str__(self) -> str:
-        return f"{self.as_f64_c():.{self._price.precision}f}"
+        return f"{self.as_f64_c():.{self._mem.precision}f}"
 
     def __repr__(self) -> str:
         return f"{type(self).__name__}('{self}')"
 
     def __del__(self) -> None:
         # https://cython.readthedocs.io/en/latest/src/userguide/special_methods.html#finalization-methods-dealloc-and-del
-        price_free(self._price)  # `self._price` moved to Rust (then dropped)
+        price_free(self._mem)  # `self._mem` moved to Rust (then dropped)
 
     cdef bint eq(self, Price other) except *:
-        return self._price.fixed == other._price.fixed
+        return self._mem.raw == other._mem.raw
 
     cdef bint ne(self, Price other) except *:
-        return self._price.fixed != other._price.fixed
+        return self._mem.raw != other._mem.raw
 
     cdef bint lt(self, Price other) except *:
-        return self._price.fixed < other._price.fixed
+        return self._mem.raw < other._mem.raw
 
     cdef bint le(self, Price other) except *:
-        return self._price.fixed <= other._price.fixed
+        return self._mem.raw <= other._mem.raw
 
     cdef bint gt(self, Price other) except *:
-        return self._price.fixed > other._price.fixed
+        return self._mem.raw > other._mem.raw
 
     cdef bint ge(self, Price other) except *:
-        return self._price.fixed >= other._price.fixed
+        return self._mem.raw >= other._mem.raw
 
     cdef bint is_zero(self) except *:
-        return self._price.fixed == 0
+        return self._mem.raw == 0
 
     cdef bint is_negative(self) except *:
-        return self._price.fixed < 0
+        return self._mem.raw < 0
 
     cdef bint is_positive(self) except *:
-        return self._price.fixed > 0
+        return self._mem.raw > 0
 
     @staticmethod
-    def from_fixed(int64_t fixed, uint8_t precision):
-        return Price.from_fixed_c(fixed, precision)
+    def from_raw(int64_t raw, uint8_t precision):
+        return Price.from_raw_c(raw, precision)
 
     @staticmethod
-    cdef Price from_fixed_c(int64_t fixed, uint8_t precision):
+    cdef Price from_raw_c(int64_t raw, uint8_t precision):
         cdef Price price = Price.__new__(Price)
-        price._price = price_from_fixed(fixed, precision)
+        price._mem = price_from_raw(raw, precision)
         return price
 
-    cdef int64_t fixed_int64_c(self) except *:
-        return self._price.fixed
+    cdef int64_t raw_int64_c(self) except *:
+        return self._mem.raw
 
     cdef double as_f64_c(self) except *:
-        return self._price.fixed / FIXED_SCALAR
+        return self._mem.raw / FIXED_SCALAR
 
     @staticmethod
     cdef object _extract_decimal(object obj):
@@ -619,7 +619,7 @@ cdef class Price:
 
     @property
     def precision(self) -> int:
-        return self._price.precision
+        return self._mem.precision
 
     @staticmethod
     cdef Price from_str_c(str value):
@@ -690,7 +690,7 @@ cdef class Price:
         Decimal
 
         """
-        return decimal.Decimal(f"{self.as_double():.{self._price.precision}f}")
+        return decimal.Decimal(f"{self.as_double():.{self._mem.precision}f}")
 
     cpdef double as_double(self) except *:
         """
@@ -704,18 +704,18 @@ cdef class Price:
         return self.as_f64_c()
 
     cdef Price add(self, Price other):
-        cdef int64_t fixed = self._price.fixed + other.fixed_int64_c()
-        return Price.from_fixed_c(fixed, self._price.precision)
+        cdef int64_t raw = self._mem.raw + other.raw_int64_c()
+        return Price.from_raw_c(raw, self._mem.precision)
 
     cdef Price sub(self, Price other):
-        cdef int64_t fixed = self._price.fixed - other.fixed_int64_c()
-        return Price.from_fixed_c(fixed, self._price.precision)
+        cdef int64_t raw = self._mem.raw - other.raw_int64_c()
+        return Price.from_raw_c(raw, self._mem.precision)
 
     cdef void add_assign(self, Price other) except *:
-        self._price.fixed += other.fixed_int64_c()
+        self._mem.raw += other.raw_int64_c()
 
     cdef void sub_assign(self, Price other) except *:
-        self._price.fixed -= other.fixed_int64_c()
+        self._mem.raw -= other.raw_int64_c()
 
 
 cdef class Money:
@@ -734,36 +734,36 @@ cdef class Money:
         if value is None:
             value = 0
 
-        self._money = money_new(float(value), <Currency_t>currency._currency)  # borrows wrapped `currency`
+        self._mem = money_new(float(value), <Currency_t>currency._currency)  # borrows wrapped `currency`
         self.currency = currency
 
     def __getstate__(self):
-        return self._money.fixed, self.currency
+        return self._mem.raw, self.currency
 
     def __setstate__(self, state):
         cdef Currency currency = state[1]
-        self._money = money_from_fixed(state[0], <Currency_t>currency._currency)
+        self._mem = money_from_raw(state[0], <Currency_t>currency._currency)
         self.currency = currency
 
     def __eq__(self, Money other) -> bool:
         Condition.equal(self.currency, other.currency, "currency", "other.currency")
-        return self._money.fixed == other.fixed_int64_c()
+        return self._mem.raw == other.raw_int64_c()
 
     def __lt__(self, Money other) -> bool:
         Condition.equal(self.currency, other.currency, "currency", "other.currency")
-        return self._money.fixed < other.fixed_int64_c()
+        return self._mem.raw < other.raw_int64_c()
 
     def __le__(self, Money other) -> bool:
         Condition.equal(self.currency, other.currency, "currency", "other.currency")
-        return self._money.fixed <= other.fixed_int64_c()
+        return self._mem.raw <= other.raw_int64_c()
 
     def __gt__(self, Money other) -> bool:
         Condition.equal(self.currency, other.currency, "currency", "other.currency")
-        return self._money.fixed > other.fixed_int64_c()
+        return self._mem.raw > other.raw_int64_c()
 
     def __ge__(self, Money other) -> bool:
         Condition.equal(self.currency, other.currency, "currency", "other.currency")
-        return self._money.fixed >= other.fixed_int64_c()
+        return self._mem.raw >= other.raw_int64_c()
 
     def __add__(a, b) -> Union[decimal.Decimal, float]:
         if isinstance(a, float) or isinstance(b, float):
@@ -844,41 +844,41 @@ cdef class Money:
         return int(self.as_f64_c())
 
     def __hash__(self) -> int:
-        return hash((self._money.fixed, self.currency.code))
+        return hash((self._mem.raw, self.currency.code))
 
     def __str__(self) -> str:
-        return f"{self.as_f64_c():.{self._money.currency.precision}f}"
+        return f"{self.as_f64_c():.{self._mem.currency.precision}f}"
 
     def __repr__(self) -> str:
         return f"{type(self).__name__}('{str(self)}', {self.currency.code})"
 
     def __del__(self) -> None:
         # https://cython.readthedocs.io/en/latest/src/userguide/special_methods.html#finalization-methods-dealloc-and-del
-        money_free(self._money)  # `self._money` moved to Rust (then dropped)
+        money_free(self._mem)  # `self._mem` moved to Rust (then dropped)
 
     cdef bint is_zero(self) except *:
-        return self._money.fixed == 0
+        return self._mem.raw == 0
 
     cdef bint is_negative(self) except *:
-        return self._money.fixed < 0
+        return self._mem.raw < 0
 
     cdef bint is_positive(self) except *:
-        return self._money.fixed > 0
+        return self._mem.raw > 0
 
-    cdef int64_t fixed_int64_c(self):
-        return self._money.fixed
+    cdef int64_t raw_int64_c(self):
+        return self._mem.raw
 
     cdef double as_f64_c(self):
-        return self._money.fixed / FIXED_SCALAR
+        return self._mem.raw / FIXED_SCALAR
 
     @staticmethod
-    def from_fixed(uint64_t fixed, uint8_t precision):
-        return Money.from_fixed_c(fixed, precision)
+    def from_raw(uint64_t raw, uint8_t precision):
+        return Money.from_raw_c(raw, precision)
 
     @staticmethod
-    cdef Money from_fixed_c(uint64_t fixed, Currency currency):
+    cdef Money from_raw_c(uint64_t raw, Currency currency):
         cdef Money money = Money.__new__(Money)
-        money._money = money_from_fixed(fixed, <Currency_t>currency._currency)
+        money._mem = money_from_raw(raw, <Currency_t>currency._currency)
         money.currency = currency
         return money
 
@@ -901,21 +901,21 @@ cdef class Money:
 
     cdef Money add(self, Money other):
         assert self.currency == other.currency, "other money currency was not equal"  # design-time check
-        cdef int64_t fixed = self._money.fixed + other.fixed_int64_c()
-        return Money.from_fixed_c(fixed, self.currency)
+        cdef int64_t raw = self._mem.raw + other.raw_int64_c()
+        return Money.from_raw_c(raw, self.currency)
 
     cdef Money sub(self, Money other):
         assert self.currency == other.currency, "other money currency was not equal"  # design-time check
-        cdef int64_t fixed = self._money.fixed - other.fixed_int64_c()
-        return Money.from_fixed_c(fixed, self.currency)
+        cdef int64_t raw = self._mem.raw - other.raw_int64_c()
+        return Money.from_raw_c(raw, self.currency)
 
     cdef void add_assign(self, Money other) except *:
         assert self.currency == other.currency, "other money currency was not equal"  # design-time check
-        self._money.fixed += other.fixed_int64_c()
+        self._mem.raw += other.raw_int64_c()
 
     cdef void sub_assign(self, Money other) except *:
         assert self.currency == other.currency, "other money currency was not equal"  # design-time check
-        self._money.fixed -= other.fixed_int64_c()
+        self._mem.raw -= other.raw_int64_c()
 
     cpdef object as_decimal(self):
         """
@@ -926,7 +926,7 @@ cdef class Money:
         Decimal
 
         """
-        return decimal.Decimal(f"{self.as_double():.{self._money.currency.precision}f}")
+        return decimal.Decimal(f"{self.as_double():.{self._mem.currency.precision}f}")
 
     cpdef double as_double(self) except *:
         """
@@ -982,7 +982,7 @@ cdef class Money:
         str
 
         """
-        return f"{self.as_f64_c():,.{self._money.currency.precision}f} {self.currency.code}".replace(",", "_")
+        return f"{self.as_f64_c():,.{self._mem.currency.precision}f} {self.currency.code}".replace(",", "_")
 
 
 cdef class AccountBalance:
@@ -1016,10 +1016,10 @@ cdef class AccountBalance:
     ):
         Condition.equal(total.currency, locked.currency, "total.currency", "locked.currency")
         Condition.equal(total.currency, free.currency, "total.currency", "free.currency")
-        Condition.true(total.fixed_int64_c() >= 0, "total was negative")
-        Condition.true(locked.fixed_int64_c() >= 0, "locked was negative")
-        Condition.true(free.fixed_int64_c() >= 0, "free was negative")
-        Condition.true(total.fixed_int64_c() - locked.fixed_int64_c() == free.fixed_int64_c(), "total - locked != free")
+        Condition.true(total.raw_int64_c() >= 0, "total was negative")
+        Condition.true(locked.raw_int64_c() >= 0, "locked was negative")
+        Condition.true(free.raw_int64_c() >= 0, "free was negative")
+        Condition.true(total.raw_int64_c() - locked.raw_int64_c() == free.raw_int64_c(), "total - locked != free")
 
         self.total = total
         self.locked = locked
@@ -1109,8 +1109,8 @@ cdef class MarginBalance:
         InstrumentId instrument_id=None,
     ):
         Condition.equal(initial.currency, maintenance.currency, "initial.currency", "maintenance.currency")
-        Condition.true(initial.fixed_int64_c() >= 0, "initial margin was negative")
-        Condition.true(maintenance.fixed_int64_c() >= 0, "maintenance margin was negative")
+        Condition.true(initial.raw_int64_c() >= 0, "initial margin was negative")
+        Condition.true(maintenance.raw_int64_c() >= 0, "maintenance margin was negative")
 
         self.initial = initial
         self.maintenance = maintenance