+++ nautilus_trader/risk/engine.pyx
@@ -680,7 +680,7 @@ cdef class RiskEngine(Component):
             # Check failed
             return f"price {price} invalid (precision {price.precision} > {instrument.price_precision})"
         if instrument.asset_type != AssetType.OPTION:
-            if price.fixed_int64_c() <= 0:
+            if price.raw_int64_c() <= 0:
                 # Check failed
                 return f"price {price} invalid (not positive)"